requirements.txt
accelerate==0.32.1
aiohttp==3.9.5
aiosignal==1.3.1
apertium==0.2.4
apertium-streamparser==5.0.2
arrow==1.3.0
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
codecarbon==2.5.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
datasets==2.20.0
dill==0.3.8
docker-pycreds==0.4.0
evaluate==0.4.2
fairseq2==0.2.1
fairseq2n==0.2.1
fasttext==0.9.3
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.23.4
idna==3.7
Jinja2==3.1.4
jiwer==3.0.4
joblib==1.4.2
kiwisolver==1.4.5
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.1
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
overrides==7.7.0
packaging==23.2
pandas==2.2.2
peft==0.11.1
pillow==10.4.0
platformdirs==4.2.2
portalocker==2.10.0
prometheus_client==0.20.0
protobuf==5.27.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pybind11==2.13.1
pycparser==2.22
pycryptodomex==3.20.0
pynvml==11.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzipper==0.3.6
rapidfuzz==3.9.4
regex==2024.5.15
requests==2.32.3
sacrebleu==2.4.2
sacremoses==0.1.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentencepiece==0.2.0
sentry-sdk==2.8.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sonar-space==0.2.1
soundfile==0.12.1
sox==1.5.0
sympy==1.12.1
tabulate==0.9.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.2.2
torcheval==0.0.7
tqdm==4.66.4
transformers==4.42.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.17.4
xxhash==3.4.1
yarl==1.9.4
